# components/endpoints/openapi-run.yaml
swagger: '2.0'
info:
  title: Antidote Endpoints
  description: Endpoints for Antidote, Distruptive Behavior Mitigation Framework
  version: 1.0.0
host: api-endpoint-j5py2mtx7a-uc.a.run.app
schemes:
  - https
produces:
  - application/json
x-google-backend:
  address: https://api-backend-j5py2mtx7a-uc.a.run.app
  protocol: h2
paths:
  /test:
    get:
      summary: Sample response used for testing
      operationId: testcall
      responses:
        '200':
          description: Success
        '400':
          description: Bad request. User ID must be an integer and larger than 0.
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A user with the specified ID was not found.
        default:
          description: Unexpected error
  /audio:
    post:
      summary: Send audio files to Antidote
      operationId: audioInput
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: payload
          description: This endpoint is used to send audio files (as a URI) into the Antidote framework.
          schema:
            type: object
            required:
              - audio_uri
              - username
            properties:
              audio_uri:
                type: string
                description: Audio URI pointing to the location where the audio file is stored.
              username:
                type: string
                description: Unique username or user id.
      responses:
        '201':
          description: Success
        '400':
          description: Bad request. User ID must be an integer and larger than 0.
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A user with the specified ID was not found.
        default:
          description: Unexpected error
  /chat:
    post:
      summary: Send text-based chat messages to Antidote
      operationId: chatInput
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: payload
          description: This endpoint is used to send text-based chat messages into the Antidote framework.
          schema:
            type: object
            required:
              - text
              - username              
            properties:
              text:
                type: string
                description: Text-based chat message.
              username:
                type: string
                description: Unique username or user id.
              timestamp:
                type: number
                description: Unix timestamp.
            example:
              text:     'This is a test text message'
              username: 'testuser123'
      responses:
        '201':
          description: Success
        '400':
          description: Bad request. User ID must be an integer and larger than 0.
        '401':
          description: Authorization information is missing or invalid.
        '404':
          description: A user with the specified ID was not found.
        default:
          description: Unexpected error